fix: update ui

+++ CleanArchMono/data/src/main/java/com/irfanirawansukirman/data/network/model/MoviesResponse.kt
@@ -47,7 +47,7 @@ data class MoviesResult(
     @SerialName("video")
     val video: Boolean?,
     @SerialName("vote_average")
-    val voteAverage: Int?,
+    val voteAverage: Double?,
     @SerialName("vote_count")
     val voteCount: Int?
 )
